requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
Pillow
deepface
pytesseract
fpdf
jinja2